return back GridOverlays file to persist changes history
