+++ scripts/x-data-grid-pro.exports.json
@@ -386,7 +386,6 @@
   { "name": "gridNumberComparator", "kind": "Variable" },
   { "name": "GridOverlay", "kind": "Variable" },
   { "name": "GridOverlayProps", "kind": "TypeAlias" },
-  { "name": "GridOverlayWrapper", "kind": "Function" },
   { "name": "gridPageCountSelector", "kind": "Variable" },
   { "name": "gridPageSelector", "kind": "Variable" },
   { "name": "gridPageSizeSelector", "kind": "Variable" },
